대시보드 나의 활동 개선

+++ app/(protected)/(user)/dashboard/my-activity/_components/my-activity-card.tsx
@@ -1,13 +1,13 @@
 import { toast } from 'sonner';
 import { Heart } from 'lucide-react';
-import { useTransition } from 'react';
-
+import { MouseEventHandler, useState, useTransition } from 'react';
 import ImageCard from '@/components/image-card';
-
 import formatDateRange from '@/utils/formatDateRange';
 import { deleteActivity } from '@/app/action/activity';
-import MyActivityKebab from '@/app/(protected)/(user)/dashboard/my-activity/_components/my-activity-kebab';
 import { ActivityWithFavoriteAndCount } from '@/type';
+import { Button } from '@/components/ui/button';
+import { Dialog, DialogContent, DialogTitle } from '@/components/ui/dialog';
+import ActivityEditForm from './activity-edit-form';
 
 interface Props {
   activity: ActivityWithFavoriteAndCount;
@@ -17,8 +17,23 @@ export default function MyActivityCard({ activity }: Props) {
   const { id, title, views, maximumCount, startDate, endDate, _count, thumbnails } = activity;
   const dateRange = formatDateRange({ startDateString: startDate, endDateString: endDate });
   const [isPending, startTransition] = useTransition();
+  const [isEditModalOpen, setEditModalOpen] = useState<boolean>(false);
 
-  const handleDelete = () => {
+  const handleClickEditFormOpen: MouseEventHandler = (e) => {
+    e.preventDefault();
+
+    setEditModalOpen(true);
+  };
+
+  const handleDelete: MouseEventHandler = (e) => {
+    e.preventDefault();
+
+    // eslint-disable-next-line no-restricted-globals, no-alert
+    const isConfirmed = confirm('정말로 삭제하시겠습니까?');
+    if (!isConfirmed) {
+      return;
+    }
+
     startTransition(async () => {
       const action = await deleteActivity(id);
       if (!action.success) {
@@ -30,27 +45,52 @@ export default function MyActivityCard({ activity }: Props) {
   };
 
   return (
-    <div className="relative">
-      <ImageCard
-        imageSrc={thumbnails[0]}
-        isPending={isPending}
-        activityId={id}
-        title={title}
-        date={dateRange}
-        participants={maximumCount}
-        bottomContent={
-          <div className="absolute flex gap-2 text-xs bottom-3 right-3">
-            <div className="flex items-center gap-1">
-              <Heart size={15} className="text-rose-500" />
-              {_count.favorites}
+    <>
+      <div className="relative">
+        <ImageCard
+          imageSrc={thumbnails[0]}
+          isPending={isPending}
+          activityId={id}
+          title={title}
+          date={dateRange}
+          participants={maximumCount}
+          bottomContent={
+            <div className="flex flex-col w-full gap-1">
+              <div className="flex items-center justify-end w-full gap-2 text-xs text-gray-500">
+                <div className="flex items-center gap-1">
+                  <Heart size={15} className="text-red" />
+                  {_count.favorites}
+                </div>
+                <p>조회수 {views}</p>
+              </div>
+              <div className="flex w-full gap-2">
+                <Button
+                  disabled={isPending}
+                  type="button"
+                  className="w-full text-base font-medium text-white shadow"
+                  onClick={handleClickEditFormOpen}
+                >
+                  수정
+                </Button>
+                <Button
+                  disabled={isPending}
+                  type="button"
+                  className="w-full text-base font-medium text-white shadow bg-red hover:bg-rose-800"
+                  onClick={handleDelete}
+                >
+                  삭제
+                </Button>
+              </div>
             </div>
-            <p>조회수 {views}</p>
-          </div>
-        }
-      />
-      <div className="absolute top-1 right-1">
-        <MyActivityKebab handleDelete={handleDelete} activity={activity} />
+          }
+        />
       </div>
-    </div>
+      <Dialog open={isEditModalOpen} onOpenChange={setEditModalOpen}>
+        <DialogContent className="h-screen overflow-y-auto bg-white">
+          <DialogTitle className="text-2xl font-semibold">길라 활동 수정</DialogTitle>
+          <ActivityEditForm activity={activity} />
+        </DialogContent>
+      </Dialog>
+    </>
   );
 }
